+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [0.2.1] - 2020-MM-DD
 
 * Fixed comparison for intervals with other comparables, such as DateTime
+* Made IntervalTree default constructor, add(), and remove() throw an
+  ArgumentError when passing an invalid interval iterable
 
 ## [0.2.0+2] - 2020-04-27